refactor(pgmq-rs): consolidate diesel adapters under one umbrella module
Merge diesel_async and diesel_sync into adapters/diesel/, with the
diesel-async impl at the top level and diesel-sync as a sub-module.
Shared QueryableByName DTOs (SendCol, MessageRowJson, etc.) now live
once in the umbrella mod.rs instead of being duplicated across both
files.

Breaking change: pgmq::adapters::diesel_async becomes pgmq::adapters::diesel,
and pgmq::adapters::diesel_sync becomes pgmq::adapters::diesel::sync.
