+++ pgmq-rs/src/pg_ext/mod.rs
@@ -46,8 +46,8 @@
 //!
 //! - [`crate::adapters::sqlx`] — `tx.send(...)`
 //! - [`crate::adapters::tokio_postgres`] — `tx.send(...)`
-//! - [`crate::adapters::diesel_async`] — `conn.send(...)` inside `conn.transaction(|conn| ...)`
-//! - [`crate::adapters::diesel_sync`] — `conn.send(...)` inside `conn.transaction(|conn| ...)`
+//! - [`crate::adapters::diesel`] (async) — `conn.send(...)` inside `conn.transaction(|conn| ...)`
+//! - [`crate::adapters::diesel::sync`] — `conn.send(...)` inside `conn.transaction(|conn| ...)`
 //!
 //! ## See also
 //!
